+++ client/src/extension.ts
@@ -1,6 +1,7 @@
 "use strict";
-import * as fs from 'fs';
-import * as path from 'path';
+import * as fs from "fs";
+import * as path from "path";
+import * as ts from "typescript";
 import {
     commands,
     DiagnosticCollection,
@@ -14,28 +15,30 @@ import {
     window,
     workspace,
     WorkspaceConfiguration,
-} from 'vscode';
-import { LanguageClient, LanguageClientOptions, ServerOptions, SettingMonitor, TransportKind } from 'vscode-languageclient';
-import { AddI18nLabel, AddSchemaToStore, ResetI18nStorage, SwitchToController, SwitchToView } from './commands';
-import * as file from './helpers/filehandler';
-import * as log from './helpers/logging';
-import { ModuleReferenceProvider } from './language/js/ModuleReferenceProvider';
-import { I18NCompletionItemProvider } from './language/ui5/Ui5CompletionProviders';
+} from "vscode";
+import { LanguageClient, LanguageClientOptions, ServerOptions, SettingMonitor, TransportKind } from "vscode-languageclient";
+import { AddI18nLabel, AddSchemaToStore, ResetI18nStorage, SwitchToController, SwitchToView } from "./commands";
+import * as file from "./helpers/filehandler";
+import * as log from "./helpers/logging";
+import { ModuleReferenceProvider } from "./language/js/ModuleReferenceProvider";
+import { IDCompletionProvider } from "./language/ts/IDCompletionProvider";
+import * as defprov from "./language/ui5/Ui5DefinitionProviders";
 import {
     EventCallbackDefinitionProvider,
     I18nDfinitionProvider,
     Ui5ViewDefinitionProvider,
     ViewControllerDefinitionProvider,
     ViewFragmentDefinitionProvider,
-} from './language/ui5/Ui5DefinitionProviders';
-import * as defprov from './language/ui5/Ui5DefinitionProviders';
-import { ManifestCompletionItemProvider } from './language/ui5/Ui5ManifestCompletionProviders';
-import { ManifestDiagnostics } from './language/ui5/Ui5ManifestDiagnostics';
-import { Ui5EventHandlerCodeLensProvider } from './language/ui5/Ui5TsCodeLensProviders';
-import { I18nCodeActionprovider } from './language/xml/XmlActionProviders';
-import { I18nDiagnosticProvider } from './language/xml/XmlDiagnostics';
-import { Settings } from './Settings';
-import { Ui5Extension } from './UI5Extension';
+} from "./language/ui5/Ui5DefinitionProviders";
+import { ManifestCompletionItemProvider } from "./language/ui5/Ui5ManifestCompletionProviders";
+import { ManifestDiagnostics } from "./language/ui5/Ui5ManifestDiagnostics";
+import { CallbackRenameProvider } from "./language/ui5/Ui5RenameProviders";
+import { Ui5EventHandlerCodeLensProvider } from "./language/ui5/Ui5TsCodeLensProviders";
+import { I18NCompletionItemProvider } from "./language/xml/I18nCompletionProvider";
+import { I18nCodeActionprovider } from "./language/xml/XmlActionProviders";
+import { ControllerDiagnosticsProvider, I18nDiagnosticProvider } from "./language/xml/XmlDiagnostics";
+import { Settings } from "./Settings";
+import { Ui5Extension } from "./UI5Extension";
 
 export interface IDiagnose {
     diagnosticCollection: DiagnosticCollection;
@@ -47,6 +50,8 @@ export namespace ui5tsglobal {
     export const name = "ui5-ts";
     export const core: Ui5Extension = new Ui5Extension();
     export let config: Settings = new Settings();
+
+    export let tsproxy: ts.LanguageService;
 }
 
 const ui5JsonViews: DocumentFilter = { language: "json", scheme: "file", pattern: "*.view.json" };
@@ -72,13 +77,14 @@ export async function activate(c: ExtensionContext) {
     context = c;
     ui5tsglobal.core.extensionPath = c.extensionPath;
     ui5tsglobal.core.schemaStoragePath = c.asAbsolutePath("schemastore");
+    await startTypescriptLanguageService();
     // Use the console to output diagnostic information (console.log) and errors (console.error)
     // This line of code will only be executed once when your extension is activated
     console.log("Activating UI5 extension.");
 
     // Subscribe to workspace config changed configuration
     workspace.onDidChangeConfiguration(e => onDidChangeConfiguration());
-    onDidChangeConfiguration();
+    await onDidChangeConfiguration();
 
     startXmlViewLanguageServer(context);
     // startManifestLanguageServer();
@@ -97,7 +103,9 @@ export async function activate(c: ExtensionContext) {
 
     // c.subscriptions.push(languages.registerCompletionItemProvider([ui5_xmlviews, ui5_xmlfragments], new Ui5i18nCompletionItemProvider));
 
-    const diags: IDiagnose[] = [new ManifestDiagnostics(languages.createDiagnosticCollection("json")), new I18nDiagnosticProvider(languages.createDiagnosticCollection("i18n"))];
+    const diags: IDiagnose[] = [new ManifestDiagnostics(languages.createDiagnosticCollection("json")),
+    new I18nDiagnosticProvider(languages.createDiagnosticCollection("i18n")),
+    new ControllerDiagnosticsProvider(languages.createDiagnosticCollection("xml"))];
 
     createDiagnosticSubscriptions(c, diags);
 
@@ -114,6 +122,7 @@ export async function activate(c: ExtensionContext) {
     // Completionitemproviders
     c.subscriptions.push(languages.registerCompletionItemProvider(ui5Manifest, new ManifestCompletionItemProvider()));
     c.subscriptions.push(languages.registerCompletionItemProvider(ui5Xml, new I18NCompletionItemProvider()));
+    c.subscriptions.push(languages.registerCompletionItemProvider(ui5TsControllers, new IDCompletionProvider()));
 
     // Definitionproviders
     c.subscriptions.push(languages.registerDefinitionProvider(ui5View, new ViewFragmentDefinitionProvider()));
@@ -125,15 +134,21 @@ export async function activate(c: ExtensionContext) {
     // CodeLens Providers
     c.subscriptions.push(languages.registerCodeLensProvider(ui5TsControllers, new Ui5EventHandlerCodeLensProvider()));
 
+    // Rename Providers
+    if (ui5tsglobal.config.insiders) {
+        c.subscriptions.push(languages.registerRenameProvider(ui5TsControllers, new CallbackRenameProvider()));
+    }
+
     if (ui5tsglobal.config.insiders) {
         c.subscriptions.push(languages.registerReferenceProvider(javascript, new ModuleReferenceProvider()));
     }
 }
 
-function onDidChangeConfiguration() {
-    getManifestLocation();
-    getAllNamespaceMappings();
-    ResetI18nStorage();
+async function onDidChangeConfiguration() {
+    ui5tsglobal.config = new Settings();
+    await getManifestLocation();
+    await getAllNamespaceMappings();
+    await ResetI18nStorage();
 }
 
 function createDiagnosticSubscriptions(c: ExtensionContext, diags: IDiagnose[]) {
@@ -144,7 +159,8 @@ function createDiagnosticSubscriptions(c: ExtensionContext, diags: IDiagnose[]) 
         });
         workspace.onDidOpenTextDocument(diag.diagnose.bind(diag));
     }
-    for (const otd of workspace.textDocuments)
+    const docs = workspace.textDocuments;
+    for (const otd of docs)
         for (const diag of diags)
             diag.diagnose(otd);
 }
@@ -154,14 +170,15 @@ async function getManifestLocation() {
         // Get manifest location if not set in workspace settings.
         if (!ui5tsglobal.config.manifestlocation) {
             const workspaceroot = workspace.rootPath;
-            const manifest = await workspace.findFiles("**/manifest.json", "");
+            let manifest = await workspace.findFiles("**/manifest.json", "");
+            manifest = await filterAsync(manifest, async x => await isUi5Manifest(x));
             let val: string;
             if (manifest.length < 1) {
                 window.showWarningMessage("Could not find any manifest.json in your project. Please set the path to your manifest.json file in the workspace settings.");
                 return;
             } else if (manifest.length > 1) {
-                window.showInformationMessage("Multiple manifests found");
-                val = (await window.showQuickPick(manifest.map(x => ({ label: path.relative(workspace.rootPath, x.fsPath), description: x.fsPath })) as QuickPickItem[], { ignoreFocusOut: true })).label;
+                // window.showInformationMessage("Multiple manifests found");
+                val = (await window.showQuickPick(manifest.map(x => ({ label: path.relative(workspace.rootPath, x.fsPath), description: x.fsPath })) as QuickPickItem[], { ignoreFocusOut: true, placeHolder: "Multiple manifest.json files found. Please pick which one to add to your workspace config." })).label;
                 window.showInformationMessage("Set manifest.json path to workspace settings");
             } else {
                 val = manifest[0].fsPath;
@@ -176,6 +193,27 @@ async function getManifestLocation() {
 
     } catch (error) {
         // Do nothing
+    }
+}
+
+async function filterAsync<S>(inarray: S[], callback: (value: S, index: number, array: S[]) => Promise<boolean>): Promise<S[]> {
+    const ret: S[] = [];
+    let index = 0;
+    for (const entry of inarray) {
+        if (await callback(entry, index++, inarray)) {
+            ret.push(entry);
+        }
+    }
+    return ret;
+}
+
+async function isUi5Manifest(uri: Uri): Promise<boolean> {
+    try {
+        const doc = await workspace.openTextDocument(uri);
+        const manifestobject = JSON.parse(doc.getText());
+        return manifestobject["sap.app"] !== undefined;
+    } catch (error) {
+        return false;
     }
 }
 
@@ -231,7 +269,7 @@ function startXmlViewLanguageServer(c: ExtensionContext): Promise<void> {
             // Register the server for xml decuments documents
             diagnosticCollectionName: "xmlDiagnostics",
             documentSelector: ["xml", "xsd"],
-            initializationOptions: { storagepath: c.asAbsolutePath("schemastore") } as IXmlInitOptions,
+            initializationOptions: { storagepath: c.asAbsolutePath("schemastore") },
             synchronize: {
                 // Synchronize the setting section "languageServerExample" to the server
                 configurationSection: "ui5ts",
@@ -257,9 +295,44 @@ function startXmlViewLanguageServer(c: ExtensionContext): Promise<void> {
 interface IXmlInitOptions {
     /**
      * asolute path to the folder of the xsd files
-     * 
+     *
      * @type {string}
      * @memberOf XmlInitOptions
      */
     storagepath: string;
+}
+
+async function startTypescriptLanguageService(): Promise<void> {
+    // 1. Get TsConfig
+    let options: ts.CompilerOptions;
+    try {
+        const tsconfiguri = await workspace.findFiles("**/manifest.json", undefined);
+        const tsconfig = JSON.parse((await workspace.openTextDocument(tsconfiguri[0])).getText());
+        options = ts.convertCompilerOptionsFromJson(tsconfig.compilerOptions, workspace.rootPath).options;
+    } catch (error) {
+        options = ts.getDefaultCompilerOptions();
+    }
+    // 2. Get traced files
+    const files = (await workspace.findFiles("**/*.ts", undefined)).map(x => x.fsPath);
+
+    // 3. Setup Language Service Host
+    const servicesHost: ts.LanguageServiceHost = {
+        getScriptFileNames: () => files,
+        getScriptVersion: (fileName) => files[fileName] && files[fileName].version.toString(),
+        getScriptSnapshot: (fileName) => {
+            if (!fs.existsSync(fileName)) {
+                return undefined;
+            }
+
+            return ts.ScriptSnapshot.fromString(fs.readFileSync(fileName).toString());
+        },
+        getCurrentDirectory: () => process.cwd(),
+        getCompilationSettings: () => options,
+        getDefaultLibFileName: (opt) => ts.getDefaultLibFilePath(opt),
+        fileExists: ts.sys.fileExists,
+        readFile: ts.sys.readFile,
+        readDirectory: ts.sys.readDirectory,
+    };
+
+    ui5tsglobal.tsproxy = ts.createLanguageService(servicesHost);
 }